fix(express) move content from lib to src directory
