+++ modules/express/package.json
@@ -5,7 +5,11 @@
   "repository": "phenyl-js/phenyl",
   "license": "Apache License 2.0",
   "author": "Shin Suzuki <shinout310@gmail.com>, Jesse Katsumata<jesse.katsumata@gmail.com>",
-  "files": ["lib", "LICENSE", "README.md"],
+  "files": [
+    "lib",
+    "LICENSE",
+    "README.md"
+  ],
   "main": "lib/index.js",
   "types": "lib/index.d.js",
   "scripts": {
@@ -18,7 +22,8 @@
     "@phenyl/http-rules": "~1.0.0",
     "@phenyl/interfaces": "~1.0.0",
     "@phenyl/utils": "~1.0.0",
-    "raw-body": "^2.3.2"
+    "raw-body": "^2.3.2",
+    "express": "^4.16.4"
   },
   "devDependencies": {
     "upbin": "^0.8.1"
