+++ package.json
@@ -35,6 +35,7 @@
     "typescript-eslint-parser": "^21"
   },
   "devDependencies": {
+    "@types/express": "^4.16.1",
     "@types/mocha": "^5",
     "@types/node": "~8.10",
     "mocha": "^5",
